% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodMap.R
\name{FloodMap}
\alias{FloodMap}
\title{Create a Map displaying calculated Flood Areas; previous S2_data_download, S2_data_processing and SFMI_flood_calculation function runs are necessary for this function to work}
\usage{
FloodMap(
  aoi,
  title = "",
  map_file_name = "my_flood_map",
  name_of_aoi_in_legend = "Area of Interest",
  name_of_flood_areas = "Flood Areas",
  x_axis = "Longitude",
  y_axis = "Latitude",
  caption_text = "",
  sort_by_flood_size = FALSE
)
}
\arguments{
\item{aoi}{Area of Interest as the Path to a Vector File .shp or .gpkg; Warning: Always the first Feature gets selected --> If you have multiple Geometries unionize them beforehand}

\item{title}{Define a Title for the Map as a String}

\item{map_file_name}{Optional: Define File Name for the Map as a String; Default: "my_flood_map"}

\item{name_of_aoi_in_legend}{Optional: Define how your AOI will be named in the Map as a String; Default: "Area of Interest"}

\item{name_of_flood_areas}{Optional: Define how your Flood Areas will be named in the Map as a String; Default: "Flood Areas"}

\item{x_axis}{Optional: Define how your x-Axis will be named in the Map as a String; Default: "Longitude"}

\item{y_axis}{Optional: Define how your y-Axis will be named in the Map as a String; Default: "Latitude"}

\item{caption_text}{Optional: Define a Caption for your Map as a String; A new Line can be created using "BACKSLASHn"; Default: empty}

\item{sort_by_flood_size}{Optional: If you cannot see the Flood Areas clearly, this Option sorts them by size to increase visibility, as a Boolean}
}
\value{
A Map in the Working Directory in the folder "flood_map" in different Formats for Publication
}
\description{
Create a Map displaying calculated Flood Areas; previous S2_data_download, S2_data_processing and SFMI_flood_calculation function runs are necessary for this function to work
}
\examples{

aoi <- "path/to/file.shp"
aoi <- "path/to/file.gpkg"

title <- "Floods in Bavaria, Germany 2024"

map_file_name <- "flood_map_germany"

name_of_aoi_in_legend <- "Bavaria"

name_of_flood_areas <- "Inundation Areas"

x_axis <- "X"

y_axis <- "Y"

sort_by_flood_size <- TRUE

}
