% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_data_download.R
\name{S2_data_download}
\alias{S2_data_download}
\title{Download Sentinel-2 Data for Flood Mapping via the Copernicus API}
\usage{
S2_data_download(
  username,
  password,
  start_date,
  end_date,
  aoi,
  condition,
  cloud_cover_percent = 50,
  number_of_results = 6
)
}
\arguments{
\item{username}{Your E-Mail Address from Copernicus Dataspace as a String}

\item{password}{Your Password from Copernicus Dataspace as a String}

\item{start_date}{First Date Parameter as a String; Format: "YYYY-MM-DD"}

\item{end_date}{Second Date Parameter as a String; Format: "YYYY-MM-DD"}

\item{aoi}{Area of Interest as the Path to a Vector File .shp or .gpkg}

\item{condition}{Specifies the Images Depiction, either pre_flood or flood01, flood02... flood10... as a String with NO File Extension}

\item{cloud_cover_percent}{Maximum allowed Percentage of Cloud Cover on Satellite Image as an Integer}

\item{number_of_results}{Number of Results that are returned by the Copernicus API as an Integer}
}
\value{
The unprocessed raw Sentinel-2 Bands required for Flood Calculations in the Working Directory in the folder "raw-data"
}
\description{
Download Sentinel-2 Data for Flood Mapping via the Copernicus API
}
\examples{

username <- "your-email-address'AT'mail.com"

password <- "your-password"

start_date <- "2024-12-31"

end_date <- "2025-01-28"

aoi <- "link-to-file.shp"
aoi <- "link-to-file.gpkg"

condition <- "pre_flood"
condition <- "flood_01"
condition <- "flood_02"

cloud_cover_percent <- 30

number_of_results <- 12

}
