% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_data_processing.R
\name{S2_data_processing}
\alias{S2_data_processing}
\title{Processing of Sentinel-2 Bands with Cloud Masking and AOI Cropping/Masking}
\usage{
S2_data_processing(aoi, condition)
}
\arguments{
\item{aoi}{Area of Interest as the Path to a Vector File .shp or .gpkg}

\item{condition}{Specifies the Images Depiction, either pre_flood or flood01, flood02... flood10... as a String with NO File Extension}
}
\value{
The processed Sentinel-2 bands cropped and masked to your AOI, including Cloud Masking and Mosaic Creation in the Working Directory in the folder "processed-data"
}
\description{
Processing of Sentinel-2 Bands with Cloud Masking and AOI Cropping/Masking
}
\examples{

aoi <- "link-to-file.gpkg"

condition <- "flood_01"


}
